자료형 변경

+++ src/main/java/com/joebrooks/mapshotserver/service/ChromeDriverService.java
@@ -2,6 +2,7 @@ package com.joebrooks.mapshotserver.service;
 
 import com.joebrooks.mapshotserver.component.ChromeDriverEx;
 import com.joebrooks.mapshotserver.domain.KakaoMap;
+import lombok.extern.java.Log;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Dimension;
 import org.openqa.selenium.JavascriptExecutor;
@@ -43,8 +44,8 @@ public class ChromeDriverService {
         driver.get(uri.toString());
         waiter.until(ExpectedConditions.presenceOfElementLocated(By.id("checker_true")));
 
-        int requiredWidth = (Integer)driver.executeScript("return document.body.parentNode.scrollWidth");
-        int requiredHeight = (Integer)driver.executeScript("return document.body.parentNode.scrollHeight");
+        Integer requiredWidth = ((Long)driver.executeScript("return document.body.parentNode.scrollWidth")).intValue();
+        Integer requiredHeight = ((Long)driver.executeScript("return document.body.parentNode.scrollHeight")).intValue();
 
         driver.manage().window().setSize(new Dimension(requiredWidth, requiredHeight));
         byte[] srcFile = driver.findElement(By.id("map")).getScreenshotAs(OutputType.BYTES);
